refactor(PI-8): Refactorización y actualización de servicios
Se ha movido la clase `DocumentProcessorService` y sus dependencias al nuevo espacio de nombres `AppServices.DocumentProcessing`. La clase ha sido refactorizada para usar sintaxis moderna de C#, incluyendo declaraciones de espacio de nombres a nivel de archivo. Se eliminaron las clases `ChmDocumentProcessor` y `HtmlDocumentProcessor` junto con su lógica asociada.

La clase `ServicesRegistration` se actualizó para reflejar el nuevo espacio de nombres y eliminar el registro de `ChmDocumentProcessor`. La clase `DocumentProcessorService` ahora utiliza inyección de dependencias y el método `ExecuteAsync` se ha actualizado para usar estas dependencias.

Se actualizó la clase `DocumentProcessorServiceTests` para reflejar el nuevo espacio de nombres y usar la nueva sintaxis para definir la clase y sus métodos.

Además, se añadió una nueva clase `HtmlDocumentProcessor` que implementa la interfaz `IDocumentProcessor`, con métodos para procesar archivos HTML y convertirlos a Markdown, limpiarlos y dividirlos en fragmentos más pequeños.
