+++ AppServices/DocumentProcessing/DocumentProcessorService.cs
@@ -12,7 +12,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace AppServices
+namespace AppServices.DocumentProcessing
 {
     internal class DocumentProcessorService(
         IServiceProvider _services,
